Move DatabaseConnection
